changed values from float to decimal as recommended

+++ ConstructionSpending/Models/Occupancy.cs
@@ -10,7 +10,7 @@ namespace ConstructionSpending.Models
     {
         public Time Time { get; set; } //Composite Key needed requires Fluent API
         public OccupancyType OccupancyType { get; set; } //Composite Key needed requires Fluent API
-        public float? Value { get; set; }
+        public decimal? Value { get; set; }
     }
 
     public enum OccupancyType
